2. Feature: Implement a read status for messages #2
Server Changes:
I have created a column in the Message table called readAt. This column takes a datetime value for when the user has read a received message
and a has a default value of null. I have also implemented a PUT operation in server\messenger_backend\views\api\conversations.py. This performs 2 functions based on the request given.
If the request contains a conversation ID but has a message ID of null then it retrieves a conversation from the database based on the given conversation ID and updates all the messages readAt values
(if their values are null and the messages are recieved by the user)
 to timezone.now(). If the message ID is not null then it simply retrieves the message and adds a timezone.now() value to readAt.

Client Changes:
In order to determine if messages have been read I created a thunkCreator called updateReadConvo. This performs the PUT operation that performs an update to the readAt value of a given message ID or conversation ID
as well as calling a reducer I created called clearUnreadCount. This reducer then calls a reducer function called removeUnreadConvoCountFromStore that updates the unreadCount on the client sides and renders the state
accordingly. I also created a reducer called ADD_UNREAD_COUNT that increments the unread count for a recipient user using a reducer function I created called AddUnreadConvoCountToStore. To update the unread count
based on if the recipient user is currently in the conversation where the count is being incremented, I added a if condition to the client\src\socket.js "new-message" handler. I changed the activeConversation username string to take
the username's ID instead. This allows the if condition in client\src\socket.js to determine if the count needs to be incremented or not based on if the message being retrieved is
apart of the currently active conversation. If it isn't apart of the currently active conversation then it doesn't need to be incremented and the message ID is provided to updateReadConvo to
update the message's readAt value in the database.

The unread message count badge and latest message font color dynamic change (based on if the message is unread), has also been implemented.
The count is being taken from the server side with Django's ORM (convo.messages.exclude(senderId=user_id).filter(readAt=None).count())
and being sent with the initial GET of conversations JSON response.