Implement a read status for messages #2. Code review #2.
Comment 1:

I removed the getLastReceivedMessageId function entirely. To replace its functionality with improved performance and efficiency I made changes to our conversations GET and to our reducer function removeUnreadConvoCountFromStore. For the initial GET of conversations I added a lastMessageRead  key, value pair  (stores the last sent message read by the user)  to our convo_dict. To handle dynamic client changes I added a variable and condition to removeUnreadConvoCountFromStore forEach loop. The variable initially stores the lastMessageRead.id of the convoCopy to make sure the user is shown the lastMessageRead. This variable's state only changes if one of the newMessages being returned from our PUT operation returns a message who's sender is not the other user of the conversation (only update the state of lastMessageRead if the lastMessageRead is read and if its not the other user of the conversation's message). When the loop is finished it will always return the user's last message read if it was read by the other user.

Comment 2:

I removed the styleProps const variable entirely and placed the color and font weight state conditioning inside the makeStyles function.

Comment 3:

I removed the unnecessary security in server\messenger_backend\views\api\conversation.py.

Comment 4+5:

Updated all the wrongly returned status codes to their appropriate status codes.